projects/auth-lib/src/guard/logged-in.guard.ts
import { Injectable } from '@angular/core';
import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
import { Observable, of } from 'rxjs';
import { catchError, map } from 'rxjs/operators';
import { AuthenticationService } from '../service/authentication.service';

@Injectable({
  providedIn: 'root'
})
export class LoggedInGuard implements CanActivate {

  constructor(private router: Router, private authenticationService: AuthenticationService) { }

  canActivate(
    next: ActivatedRouteSnapshot,
    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
    return this.authenticationService.getUser()
      .pipe(map((user => {
        if (!user) {
          this.router.navigate(['/login']);
        }
        return !!user;
      })))
      .pipe(catchError((err: any, caught: Observable<any>) => {
        this.router.navigate(['/login']);
        return of(false);
      }));
  }

}
